Fix command-line-window
The command-line-window mode uses the FileType vim which is trigger type
for endwise. This causes an error (i.e. BEEP) when the user presses <CR>
while in insert mode in the command-line-window. It also causes problems
with search and replace (substitute) as described in issue #31.

This fixes issue #31 and the annoying **BEEP**!